Fix readme: normally_open_transactions doesn't get set
Hi,

The update to 0.7 caused our tests to stop passing.

It turned out that 655b4cd98a79ea288791be4eb9f3dd376bfd7e95 changed the behaviour of `normally_open_transactions=` and it cannot be used correctly in `test.rb` anymore. 

We managed to fix this in our project, by assigning the correct value in the RSpec hook. Would like to share the solution as others might experience it.

Best,
Adam